% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_grid.R
\name{calculate_grid}
\alias{calculate_grid}
\title{Function \code{calculate_grid}}
\usage{
calculate_grid(data, weights = NULL)
}
\arguments{
\item{data}{This is input data which is a \code{sf} object (i.e. has geometry features).}

\item{weight}{The variable to adjust the weight while assigning statistical information to the geographic hexagonal map.}
}
\value{
A geographic hexagonal map.
}
\description{
This function provides the number of hexagons for the geographical area.
}
