% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellsize.R
\name{cellsize}
\alias{cellsize}
\title{Function \code{cellsize}}
\usage{
cellsize(data, pt, nhex)
}
\arguments{
\item{data}{This is input data which is a \code{sf} object (i.e. has geometry features).}

\item{pt}{The minimum boundaries (both vertical and horizontal ones) of the geographical area.}

\item{nhex}{The number of hexagons wanted/appointed by the users (for a specific region/area).}
}
\value{
A numeric number to determine the proper \code{cellsize} to generate the suitable number of hexagons for mapping.
}
\description{
This function provides a proper \code{cellsize} for calculation of the number of hexagons.
}
